chat navigation for webinar video added

+++ modules/domain/entity/src/main/java/com/doneit/ascend/domain/entity/dto/WebinarCredentialsDTO.kt
@@ -3,5 +3,5 @@ package com.doneit.ascend.domain.entity.dto
 class WebinarCredentialsDTO(
     val key: String?,
     val link: String?,
-    val chatId: Int
+    val chatId: Long
 )
