+++ modules/features/featureMain/src/main/java/com/doneit/ascend/presentation/models/StartWebinarVideoModel.kt
@@ -4,5 +4,5 @@ import com.doneit.ascend.presentation.video_chat.states.ChatRole
 
 data class StartWebinarVideoModel(
     val role: ChatRole,
-    val chatId: Int
+    val chatId: Long
 )
